==> code/02_image_labeling.ipynb <==
import pandas as pd
import os
import shutil

df_train = pd.read_csv('../data/annotation/train_annotation.csv').drop(columns = 'Unnamed: 0')
df_test = pd.read_csv('../data/annotation/test_annotation.csv').drop(columns = 'Unnamed: 0')

## Annotation EDA

df_train.shape, df_test.shape

df_train.isnull().sum()

df_test.isnull().sum()

df_train.dtypes

df_test.dtypes

df_train['Annotation tag'].value_counts()

df_test['Annotation tag'].value_counts()

## drop goForward: no goForward in train

df_test.drop(df_test.loc[df_test['Annotation tag'] == 'goForward'].index, inplace = True)

df_test['Annotation tag'].value_counts()

## Create tag label

# new train, test folder
os.mkdir('../data/train_eda')
os.mkdir('../data/test_eda')

# train label
for i in df_train['Annotation tag'].value_counts().index:
    os.mkdir(f'../data/train_eda/{i}')

# test label 
for i in df_test['Annotation tag'].value_counts().index:
    os.mkdir(f'../data/test_eda/{i}')

def get_annotation(df,folder,annotation):
    # list_tag: store image name (no directory)
    # i is tag name: ex 'go', 'stop'
    list_tag = [i.split('/')[1] for i in df.loc[df['Annotation tag'] == annotation]['Filename']]
    list_tag_set = set(list_tag)
    list_tag = list(list_tag_set)
    print (i)
    print (list_tag[:20])
    
    # the new directory we move image into
    new_dir = f'../data/{folder}_eda/{annotation}/'
    
    # the old directory we move image from
    img_dir = f'../data/{folder}/'
    for img in list_tag:
            src = os.path.join(img_dir, img)
            dst = os.path.join(new_dir, img)
            shutil.copyfile(src,dst)

# train
for i in df_train['Annotation tag'].value_counts().index:
    get_annotation(df_train, 'train', i)

# test
for i in df_test['Annotation tag'].value_counts().index:
    get_annotation(df_test, 'test', i)

## remove train, test folder

shutil.rmtree('../data/train')
shutil.rmtree('../data/test')
