==> code/03_shrink_image.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
import os
from PIL import Image
import itertools

df_train = pd.read_csv('../data/annotation/train_annotation.csv').drop(columns = 'Unnamed: 0')
df_test = pd.read_csv('../data/annotation/test_annotation.csv').drop(columns = 'Unnamed: 0')

# create image name column (no folder name before)
df_train['images'] = df_train['Filename'].apply(lambda x: x.split('/')[1])
df_test['images'] = df_test['Filename'].apply(lambda x: x.split('/')[1])

df_test.head()

# new train folder
# os.mkdir('../data/new_train')

def shrink_image(annotation):
   
   
    # get old image directory with 1 annotation in a list from train_eda
    # output format: ../data/..../imagename.jpg
    direct = [f'../data/train_eda/{annotation}/{image}' \
              for image in df_train.loc[df_train['Annotation tag'] == annotation]['images']]
    direct_set = set(direct)
    direct = list(direct_set)
    
    
    # get all x,y coordinate
    # output format: dictionary  {image full directory: [coordinate, coordinate...]}
    y1 = {}
    for i in direct:
        y = list(df_train.loc[df_train['images'] == i.split('/')[4]]['Upper left corner Y'])
        y1[i] = y
    y2 = {}
    for i in direct:
        y = list(df_train.loc[df_train['images'] == i.split('/')[4]]['Lower right corner Y'])
        y2[i] = y
    x1 = {}
    for i in direct:
        x = list(df_train.loc[df_train['images'] == i.split('/')[4]]['Upper left corner X'])
        x1[i] = x
    x2 = {}
    for i in direct:
        x = list(df_train.loc[df_train['images'] == i.split('/')[4]]['Lower right corner X'])
        x2[i] = x
    
    
    # get all images in np.array
    # i is image full directory
    # i[y1:y2, x1:x2]
    imgs = []
    for i in x1:
        for j in range(len(x1[i])):
            img = plt.imread(i)[y1[i][j]:y2[i][j] , x1[i][j]:x2[i][j]]
            imgs.append(img)
    
    # create new directory
    os.mkdir(f'../data/new_train/{annotation}')
    
    # move image to directory
    for i,image in enumerate(imgs):
        im = Image.fromarray(image)
        im.save(f'../data/new_train/{annotation}/{i}.jpg')
    

shrink_image('go')

shrink_image('stop')

shrink_image('stopLeft')

shrink_image('goLeft')

shrink_image('warningLeft')

shrink_image('warning')

# new test folder
# os.mkdir('../data/new_test')

def shrink_image_test(annotation):
   
   
    # get old image directory with 1 annotation in a list from train_eda
    # output format: ../data/..../imagename.jpg
    direct = [f'../data/test_eda/{annotation}/{image}' \
              for image in df_test.loc[df_test['Annotation tag'] == annotation]['images']]
    direct_set = set(direct)
    direct = list(direct_set)
    
    
    # get all x,y coordinate
    # output format: dictionary  {image full directory: [coordinate, coordinate...]}
    y1 = {}
    for i in direct:
        y = list(df_test.loc[df_test['images'] == i.split('/')[4]]['Upper left corner Y'])
        y1[i] = y
    y2 = {}
    for i in direct:
        y = list(df_test.loc[df_test['images'] == i.split('/')[4]]['Lower right corner Y'])
        y2[i] = y
    x1 = {}
    for i in direct:
        x = list(df_test.loc[df_test['images'] == i.split('/')[4]]['Upper left corner X'])
        x1[i] = x
    x2 = {}
    for i in direct:
        x = list(df_test.loc[df_test['images'] == i.split('/')[4]]['Lower right corner X'])
        x2[i] = x
    
    
    # get all images in np.array
    # i is image full directory
    # i[y1:y2, x1:x2]
    imgs = []
    for i in x1:
        for j in range(len(x1[i])):
            img = plt.imread(i)[y1[i][j]:y2[i][j] , x1[i][j]:x2[i][j]]
            imgs.append(img)
    
    # create new directory
    os.mkdir(f'../data/new_test/{annotation}')
    
    # move image to directory
    for i,image in enumerate(imgs):
        im = Image.fromarray(image)
        im.save(f'../data/new_test/{annotation}/{i}.jpg')
    

shrink_image_test('go')

# kills the kernal
#shrink_image_test('stop')

shrink_image_test('stopLeft')

shrink_image_test('goLeft')

shrink_image_test('warningLeft')

shrink_image_test('warning')

# kills the kernal
#shrink_image_test('stop')

shrink_image_test('stopLeft')

shrink_image_test('goLeft')

shrink_image_test('warningLeft')

shrink_image_test('warning')

# for stop in test
# get old image directory with 1 annotation in a list from train_eda
# output format: ../data/..../imagename.jpg
direct = [f'../data/test_eda/stop/{image}' \
          for image in df_test.loc[df_test['Annotation tag'] == 'stop']['images']]
direct_set = set(direct)
direct = list(direct_set)
len(direct)

# get all x,y coordinate
# output format: dictionary  {image full directory: [coordinate, coordinate...]}
y1 = {}
for i in direct:
    y = list(df_test.loc[df_test['images'] == i.split('/')[4]]['Upper left corner Y'])
    y1[i] = y

y2 = {}
for i in direct:
    y = list(df_test.loc[df_test['images'] == i.split('/')[4]]['Lower right corner Y'])
    y2[i] = y

x1 = {}
for i in direct:
    x = list(df_test.loc[df_test['images'] == i.split('/')[4]]['Upper left corner X'])
    x1[i] = x

x2 = {}
for i in direct:
    x = list(df_test.loc[df_test['images'] == i.split('/')[4]]['Lower right corner X'])
    x2[i] = x

len(y1),len(y2)

len(x1),len(x2)

# create new directory
#os.mkdir(f'../data/new_test/stop')


# get all images in np.array
# separate stop into 2 groups
out1 = dict(itertools.islice(x1.items(), 5000))
out2 = dict(list(x1.items())[5001: 6000])
out3 = dict(list(x1.items())[6001: 7000])
out4 = dict(list(x1.items())[7001: 8000])
# out5 = dict(list(x1.items())[8001: 9000])        kills kernal
# out6 = dict(list(x1.items())[9001: 9943])        kills kernal

# i is image full directory
# i[y1:y2, x1:x2]
# import first 5000
imgs = []
for i in out1:
    for j in range(len(out1[i])):
        img = plt.imread(i)[y1[i][j]:y2[i][j] , x1[i][j]:x2[i][j]]
        imgs.append(img)

for i,image in enumerate(imgs):
    im = Image.fromarray(image)
    im.save(f'../data/new_test/stop/{i}_1.jpg')

imgs = []
for i in out2:
    for j in range(len(out2[i])):
        img = plt.imread(i)[y1[i][j]:y2[i][j] , x1[i][j]:x2[i][j]]
        imgs.append(img)

for i,image in enumerate(imgs):
    im = Image.fromarray(image)
    im.save(f'../data/new_test/stop/{i}_2.jpg')

imgs = []
for i in out3:
    for j in range(len(out3[i])):
        img = plt.imread(i)[y1[i][j]:y2[i][j] , x1[i][j]:x2[i][j]]
        imgs.append(img)

for i,image in enumerate(imgs):
    im = Image.fromarray(image)
    im.save(f'../data/new_test/stop/{i}_3.jpg')

imgs = []
for i in out4:
    for j in range(len(out4[i])):
        img = plt.imread(i)[y1[i][j]:y2[i][j] , x1[i][j]:x2[i][j]]
        imgs.append(img)

for i,image in enumerate(imgs):
    im = Image.fromarray(image)
    im.save(f'../data/new_test/stop/{i}_4.jpg')

out5 = dict(list(x1.items())[8001: 9000])
out5_1 = dict(list(x1.items())[8001: 8501])
out5_2 = dict(list(x1.items())[8501: 9001])


imgs = []
for i in out5_1:
    for j in range(len(out5_1[i])):
        img = plt.imread(i)[y1[i][j]:y2[i][j] , x1[i][j]:x2[i][j]]
        imgs.append(img)
for i,image in enumerate(imgs):
    im = Image.fromarray(image)
    im.save(f'../data/new_test/stop/{i}_5_1.jpg')

imgs = []
for i in out5_2:
    for j in range(len(out5_2[i])):
        img = plt.imread(i)[y1[i][j]:y2[i][j] , x1[i][j]:x2[i][j]]
        imgs.append(img)
for i,image in enumerate(imgs):
    im = Image.fromarray(image)
    im.save(f'../data/new_test/stop/{i}_5_2.jpg')

out6 = dict(list(x1.items())[9001: 9943])
out6_1 = dict(list(x1.items())[9001: 9501])
out6_2 = dict(list(x1.items())[9501: 9944])

for i in out6_1:
    for j in range(len(out6_1[i])):
        img = plt.imread(i)[y1[i][j]:y2[i][j] , x1[i][j]:x2[i][j]]
        imgs.append(img)

for i,image in enumerate(imgs):
    im = Image.fromarray(image)
    im.save(f'../data/new_test/stop/{i}_6_1.jpg')

for i in out6_2:
    for j in range(len(out6_2[i])):
        img = plt.imread(i)[y1[i][j]:y2[i][j] , x1[i][j]:x2[i][j]]
        imgs.append(img)

for i,image in enumerate(imgs):
    im = Image.fromarray(image)
    im.save(f'../data/new_test/stop/{i}_6_2.jpg')
